Fancier Markdown tester
A few changes to make our Markdown test runner more useful:

1. `cd`s into the directory with the `.md` file before executing the
tests. This will let us read files in them. There's a
`contextlib.chdir()` in Python 3.11, but since we want to support
before that, copies it into here.

2. More robust output capturing. `contextlib.redirect_stdout()`
doesn't work when subprocesses are spawned and a few other
cases. Steals this functionality from the internals of pytest via
`_pytest.capture.FDCapture()`.

3. Captures exceptions during compilation as part of the "found
output". Not like we want to test for `SyntaxError`s, but it makes
them easier to read.

4. Refactor into `TestCode` and `ExpectedOutput` classes. This will
make it more explicit what is coming from where.

5. Support `doctest:XXX` flags which will change how the output is
checked against the expected output blocks. We read the flags from the
"info line" and store these flags on the `TestCode`, then make our own
version of the doctest checker object `MdOutputChecker` which applies
the transformations on the found or expected output. Added
`IGNORE_OUTPUT`, `SORT_OUTPUT`, and `SORT_EXPECTED` flags. You can
read about them in the docstring.