Don't use PY_IGNORE_MISMATCH or run doctests in CI
pytest's doctest runner walks a Python source tree, tries importing
the modules, then looks at the docstrings of each object and tries
running them. This means in CI, if we build and install the `bytewax`
library, when pytest walks the source tree, the code it imports isn't
the file it's looking at and errors out. See
https://github.com/pytest-dev/pytest/issues/2042#issuecomment-381309723

One way to "fix" that is to specify `PY_IGNORE_IMPORTMISMATCH=1` on
the command line. But as far as I can tell, this means that the import
process used by `multiprocessing` to start up subprocesses no longer
works (even outside of doctests!) and we error out with "can't import
the PyO3 `.so` file" errors. I think it's because the subprocess
starts up and tries to import the working directory `bytewax` instead
of the installed one. I'm not totally sure here, but in any case it's
not working.

Turn off running doctests in CI until we can figure this out. My long
term plan is to write a new doctest discover-er that works with
imported packages. The idea would be to just read the docstrings and
run them similar to the markdown tests. (Currently doctests import the
surrounding code, which is what makes for all the import trouble, so
we'd be loosing that, but I think that's fine.)