Go back to `redirect_stdout()` in Markdown test runner
Do this because Timely logs random warnings to stdout occasionaly and
we can't turn them off. `redirect_stdout()` only captures _Python_
output, so we can hack-ily take advantage of this to only get the
output we actually care about in tests. It would be better if Timely
always logged in a way we could configure.