Automatic state recovery / garbage collection
This is an overhaul of the state recovery system to support two big
new features:

1. _Automatic recovery_ - Bytewax will backup in the recovery store
the **dataflow frontier** which is the oldest in-progress epoch. This
represents the epoch that if the dataflow crashed right now, one
should resume from. The startup machinery now reads this epoch from
the recovery store and will pass it to the input builders so the
system will start up exactly where it failed.

2. _State garbage collection_ - By keeping track of the dataflow
frontier, we can can know what state is no longer necessary for
recovery and GC it from the state store.

See the module docstring in `bytewax.recovery` for what is necessary
from a user's perspective and see the rustdoc in the `recovery` module
to see documentation on how this works internally.