Only snapshot for recovery once per state key
We only need to snapshot the logic once per state key. Stores
activated state keys in a `HashSet` to do this. Previously, we were
building up a `Vec`, pushing the state key onto the end with each
input data with that state key; this means we would snapshot once for
each input item.

Takes a look at all of the persistent data structures as part of the
`StatefulUnary` operator and changes them to be their most semantic
versions. Hopefully that'll prevent other bugs like this here:

- `activate_epochs_buffer` to a sorted set (we should activate the
  epochs in order and just once).

- `keys_buffer` to a set (we just need one copy of each key).

- `activated_epochs_to_keys_buffer` to a map to a set (the original
  problem).