Eagerly build logic and assert built
Currently, we're slowly re-building window and stateful logic structs
after recovery as we re-encounter those keys (state or window key)
lazily. That's fine, but means that everywhere we might need to access
the logic we need to do the lazy building and in at least one
instance, I forgot to do that: when a window is closed.

This resulted in a bug I found in a different PR: We create a window
for a key. Crash. Recover. Then the next thing is that the window
closes, without an input item for that window being processed. Since
we re-built the list of currently open windows from the recovery data,
we know we need to close it, but we weren't asserting that the logic
acutally exists and only ran the close code if it did.

This change asserts that the logic always exists whenever some other
state detects it should.

Also to pre-empt bugs of this type, changes to eagerly re-building
logic from the recovery data on restart, rather than lazily as input
comes in. This might change the speed of recovery restarts, but you'll
be paying that time anyway at some point, so I'm not sure that's too
worrying.

Tests to make sure this bug is gone will be forthcoming in the merge
of the fate PR.