Adds collect window operator
Adds a new windowing operator that returns a time-sorted list of all
the values within a window.

Tests.

Had to thread through the `item_time` into the
`WindowLogic::with_next` to allow consistent sorting. We might want to
expose that to the custom `fold_window` and `reduce_window` operators
as of now they don't guarantee any order of items being aggregated.