Do not use built-in `hash` for routing
We totally forgot that `hash` is _not_ consistent across Python
process invocations. Thus if we use it for output routing, we won't be
consistently routing data into partitions which breaks the promise of
partitioned output.

Currently our only partitioned output with more than one partition is
`DirOutput`, so makes that default to using `zlib.adler32`, which is a
simple consistent hash function. This will not prevent hash-based DoS,
but if you're concerned about that, you need to take a wholistic
approach to prevention anyway.

Adds a comment to `PartitionedOutput.assign_part` to describe the
dangers of `hash`. This uses `pdoc`'s support for RST admonitions.

Also adds a slightly more descriptive panic message when routing is
inconsistent.

Also changes to calculate the partition key once in a `map` operator
and re-use that, rather than calculate it twice.