AsyncBatcher
Another tool for fitting async input sources into our cooperative
multitasking world.

If you have an input source that is blocking, but has a timeout param,
great! You can use that to make a well-behaved source `next` function
that returns a single item. If you want to return a batch of items,
hopefully your input source has a batch size param.

If you don't have a batch size param, though, things get tricky. You
either have poll individual items via a blocking API and pick a
timeout for each item, then keep track of the total time spent so you
yield manually to not hog time. Or if you have an async API, you have
to wrangle with async timeouts and cancellation.

This PR adds a utility class `AsyncBatcher` which lets you say "wait
up to X timeout or N items and gather them all into a batch" when the
source is an async iterator. This lets you more easily play nicely
with coop multitasking but still use async APIs.

You can use this with sync+timeout APIs too, but you'll have to
artificially add in a `await asyncio.sleep(0)` in the "hot loop" to
provide a cancellation point.

Updates the wikistream example. And actually the wikistream example
was always "bad" because the previous sync SSE client was blocking
without any timeouts. Naughty. This converts it to use an async client
so that even if no items come in, we yield from the input operator
every 1/4th sec.

This is sort of complementary to the `awake_next` work: this helps you
manage the work the source is doing while it is awake, but it has no
affect on when it is awoken. Again, the holy grail is to tie the
Python run loop to a Timely `Activator` so that there's no "awake
next" time, it's just whenever the Python async runtime knows we have
more data.